build: 完成Qt6.8迁移

+++ apps/gui/browser_startup_settings_dialog.cpp
@@ -133,8 +133,9 @@ void BrowserStartupSettingsDialog::setCurrentTabUrls(const QStringList& urls,
     currentTabUrls_ = urls;
     currentTabIndex_ = currentTabUrls_.isEmpty()
                            ? 0
-                           : std::clamp(currentIndex, 0,
-                                        currentTabUrls_.size() - 1);
+                           : std::clamp(
+                                 currentIndex, 0,
+                                 static_cast<int>(currentTabUrls_.size()) - 1);
     updateActions();
 }
